# MediClaim.API/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MediClaim API",
    "description": "Enterprise Claims API",
    "contact": {
      "name": "MediClaim Team",
      "email": "support@mediclaim.com"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:44312",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/claims/draft": {
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "createDraft",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftClaimCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftClaimCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftClaimCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/my": {
      "get": {
        "tags": [
          "Claims"
        ],
        "operationId": "getMyClaims",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/submit": {
      "post": {
        "tags": [
          "Claims"
        ],
        "summary": "Submit draft claim.",
        "description": "Transitions claim from Draft state to Submitted state.",
        "operationId": "submit",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "Claim submitted successfully."
          },
          "400": {
            "description": "Validation failure."
          },
          "404": {
            "description": "Claim not found."
          },
          "409": {
            "description": "Claim already submitted."
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/officer-queue": {
      "get": {
        "tags": [
          "Claims"
        ],
        "operationId": "getOfficerQueue",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/transition": {
      "patch": {
        "tags": [
          "Claims"
        ],
        "operationId": "transition",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionClaimCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionClaimCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionClaimCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/approve": {
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "approve",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/reject": {
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "reject",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectClaimCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectClaimCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RejectClaimCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/escalate": {
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "escalate",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/settle": {
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "settle",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/{claimId}/documents": {
      "post": {
        "tags": [
          "Claims"
        ],
        "operationId": "uploadDocument",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/v1/claims/fraud-flags": {
      "get": {
        "tags": [
          "Claims"
        ],
        "operationId": "getFraudFlags",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal server error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ClaimStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9
        ],
        "type": "integer",
        "format": "int32"
      },
      "CreateDraftClaimCommand": {
        "type": "object",
        "properties": {
          "policyId": {
            "type": "string",
            "format": "uuid"
          },
          "providerId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "diagnosisCode": {
            "type": "string",
            "nullable": true
          },
          "treatmentCategory": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RejectClaimCommand": {
        "type": "object",
        "properties": {
          "claimId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TransitionClaimCommand": {
        "type": "object",
        "properties": {
          "claimId": {
            "type": "string",
            "format": "uuid"
          },
          "newStatus": {
            "$ref": "#/components/schemas/ClaimStatus"
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}